simulation_software/requirements.txt
# 仿真软件管理系统依赖包

# GUI框架
PyQt5>=5.15.0

# 数据库连接
mysql-connector-python>=8.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# XML处理
lxml>=4.6.0

# 自动化控制（可选）
pyautogui>=0.9.53

# 图像处理（用于图像识别模块）
opencv-python>=4.5.0
Pillow>=8.0.0

# 科学计算
scipy>=1.7.0

# 机器学习（用于参数寻优模块）
scikit-learn>=1.0.0

# 绘图
matplotlib>=3.4.0
seaborn>=0.11.0

# 其他工具
pathlib2>=2.3.0
configparser>=5.0.0